bender/requirements.txt
flask
gunicorn
redis
rq
pyyaml
jira
opsgenie-sdk
